Fix usertag as prefix for Discord; Add prefix reset command
